// src/app/pages/base-list/shared/base-list-enum.ts
export enum BaseListEnum {
    hotelList = 'hotels',
    transferList = 'transfers',
    activityList = 'activities',

    // Paginator data
    length = 100,
    pageSize = 5,

    // Titles
    hotelTitle = 'Active Establishments',
    transferTitle = 'Active Transfers',
    activityTitle = 'Active Activities',

    emptyString = '',
    categoryList = 'categoryList',

    selectedEstablishmentIdCookieName = 'selectedEstablishmentId',
    selectedTransferIdCookieName = 'selectedTransferId',
}